% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{ABCcv}
\alias{ABCcv}
\title{Cross-validation}
\usage{
ABCcv(XY)
}
\arguments{
\item{XY}{One of the datasets in this packages, e.g., as loaded by data(MACCSbinary)}
}
\value{
A data frame containing cross-validation predictions
}
\description{
Evaluates a single random cross-validation split for a set of models
}
