% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{ABCpredict}
\alias{ABCpredict}
\title{Predictions on test data}
\usage{
ABCpredict(ABCmodel, XY)
}
\arguments{
\item{ABCmodel}{A model previously trained by ABCtrain()}

\item{XY}{A dataset where test data is taken to be those samples where Label is NA}
}
\value{
Test data with predictions augmented as a new ABCpred column
}
\description{
Applies a previously-trained ABC model to new data
}
